Introduced ConstraintExprs as the extracted representation of types and some basic support for subtype checking among QualifiedTypes. (Huge commit)
This affects several parts of the typer: Each Type gets a member cExpr, which is its lazily extracted ConstraintExpr. Types such as TermRef, ConstantType and QualifiedType get precise cExprs, others just get a trivial ones ("true"). This covers literal and reference primitives on the qualifier level. Furthermore, we inject qualified result types for primitive operations, such as `Int.+`. These also come with special subclasses of ConstraintExpr. Each ConstraintExpr makes the types it depends on explicit (`cExpr.scope`) and propagates mapping operations on the corresponding type across this scope (substitutions, asSeenFrom; others may be missing still).
This commit also contains the bare necessities for performing subtyping checks among qualified types: We added QTypeConstraint, which encapsulates one subtyping constraint to be checked on the cExpr-level. TypeComparer now instantiates such a QTypeConstraint if the rhs type is a QualifiedType, and subsequently solves the constraint using ConstraintChecker. All of this heavily relies on the Stainless constraint solving infrastructure, and assumes that an SMT solver is available on the system (see stainless for more information).