Large refactoring of QTypes and CExprs now that the overall architecture seems to be in place.
In particular, we returned to using QTypes as BindingTypes, along with the QualifierSubject type that is now assigned to the qualifier variable in manually ascribed QTypes. To this end QualifiedType has been split into two representations: PrimitiveQType for those types injected as result types of primitive operations, and ComplexQType, representing user-given QualifiedTypes.
This refactoring also resolves our previously dubious pickling of QualifiedTypes.

As a result, all of the existing Dotty tests (including, afaict, the boostrapping ones) now pass for the first time. The bad news is that they take a long time to run (~4x), though I have not looked into this at all so far.